+++ ratect-core/src/lib.rs
@@ -2,4 +2,5 @@ pub mod config;
 pub mod docker;
 pub mod engine;
 pub mod expressions;
+pub mod proxy;
 pub mod user;